requirements.txt
torch==2.3.1
torchvision==0.18.1
PyYAML
matplotlib
Pillow
tqdm
diffusers
transformers
accelerate
numpy